Attribute 5 now setable, IP and subnet mask changable

+++ source/src/cip/cipcommon.c
@@ -366,7 +366,7 @@ CipAttributeStruct *GetCipAttribute(const CipInstance *const instance,
                                     const EipUint16 attribute_number) {
 
   CipAttributeStruct *attribute = instance->attributes; /* init pointer to array of attributes*/
-  for (int i = 0; i < instance->cip_class->number_of_attributes; i++) {
+  for (size_t i = 0; i < instance->cip_class->number_of_attributes; i++) {
     if (attribute_number == attribute->attribute_number) {
       return attribute;
     } else {
@@ -542,7 +542,7 @@ int EncodeData(const EipUint8 cip_type,
     case (kCipUdintUdintUdintUdintUdintString): {
       /* TCP/IP attribute 5 */
       CipTcpIpNetworkInterfaceConfiguration *
-      tcp_ip_network_interface_configuration =
+        tcp_ip_network_interface_configuration =
         (CipTcpIpNetworkInterfaceConfiguration *) cip_data;
       counter += AddDintToMessage(
         ntohl(tcp_ip_network_interface_configuration->ip_address),
